Fix flash attention + bf16 dtype mismatch in LoKr/LoHa networks
Cast diff_weight to input dtype in LoKr and LoHa forward methods.
When bf16 mixed precision is used without autocast (use_fp32=False),
F.linear/F.conv2d fails due to bf16 input vs fp32 weight mismatch.

Reverts the anima_models.py change that forced use_fp32=True for bf16,
which would have unnecessarily cast all intermediate computations to
float32 and changed training behavior. Instead, fixes the root cause
at the network layer level, matching how OFT already handles this
with .to(org_dtype).

+++ networks/loha.py
@@ -253,7 +253,7 @@ class LoHaModule(torch.nn.Module):
             if torch.rand(1) < self.module_dropout:
                 return org_forwarded
 
-        diff_weight = self.get_diff_weight()
+        diff_weight = self.get_diff_weight().to(x.dtype)
 
         # rank dropout (applied on output dimension)
         if self.rank_dropout is not None and self.training:
@@ -379,7 +379,7 @@ class LoHaInfModule(LoHaModule):
         return weight
 
     def default_forward(self, x):
-        diff_weight = self.get_diff_weight()
+        diff_weight = self.get_diff_weight().to(x.dtype)
         if self.is_conv:
             if self.conv_mode == "1x1":
                 diff_weight = diff_weight.unsqueeze(2).unsqueeze(3)
